ssaes-backend/requirements.txt
amqp==5.3.1
asgiref==3.11.0
astroid==4.0.2
asttokens==3.0.1
async-timeout==5.0.1
billiard==4.2.4
black==25.12.0
celery==5.6.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
decorator==5.2.1
dill==0.4.0
Django==5.2.9
django-cors-headers==4.9.0
django-filter==25.2
django_celery_results==2.6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
exceptiongroup==1.3.1
executing==2.2.1
ipython==8.37.0
isort==7.0.0
jedi==0.19.2
joblib==1.5.3
kombu==5.6.1
lightgbm==4.6.0
matplotlib-inline==0.2.1
mccabe==0.7.0
mypy_extensions==1.1.0
numpy==2.2.6
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
parso==0.8.5
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pure_eval==0.2.3
Pygments==2.19.2
PyJWT==2.10.1
pylint==4.0.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytokens==0.3.0
pytz==2025.2
redis==7.1.0
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
sqlparse==0.5.4
stack-data==0.6.3
threadpoolctl==3.6.0
tomli==2.3.0
tomlkit==0.13.3
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.2.14
xgboost==3.1.2
